phoenix_pubsub v1.0.2 Phoenix.PubSub.GC

A garbage collector process that cleans up the table used by Phoenix.PubSub.Local.

Functions

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

down(gc_server, pid)

Force table clean up because the given pid is down asynchronously.

gc_server - The registered server name or pid

pid - The subscriber pid

Examples

iex> down(:gc_server, self())
:ok

init(arg)

start_link(server_name, local_name)

Starts the server.

server_name - The name to register the server under

table_name - The name of the local table